requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-chroma==0.1.4
chromadb>=0.4.0,!=0.5.4,!=0.5.5,<0.6.0
streamlit==1.38.0
pypdf==4.3.1
python-dotenv==1.0.1
openai==1.45.0
tiktoken==0.7.0
ollama==0.3.3
sentence-transformers==3.1.0